requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
wandb>=0.15.0
numpy>=1.24.0
trimesh>=3.21.0
open3d>=0.17.0
matplotlib>=3.7.0
pillow>=9.5.0
pyvista>=0.40.0
clip @ git+https://github.com/openai/CLIP.git
ftfy>=6.1.1
regex>=2023.5.5
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.2.0
einops>=0.6.1
timm>=0.9.0
tensorboard>=2.13.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0
sentencepiece>=0.1.99
protobuf>=3.20.0
h5py>=3.8.0
pymeshlab>=2022.2
cadquery>=2.2.0
ezdxf>=1.0.0
pyparsing>=3.0.9